Moved TritsToBytes() and BytesToTrits() to kerl package
The former implementation of TritsToBytes and BytesToTrits is only needed for hashing with kerl.
These functions were moved to the kerl package like in the "entangled common" repository.

New implementations of TritsToBytes and BytesToTrits were added, which pack and unpack 5 trits in one byte (e.g. for sending messages to IRI).
